# docs/swagger.yaml
definitions:
...
  controllers.roleInput:
    properties:
      name:
        type: string
    type: object
...
  models.Role:
    properties:
      id:
        type: integer
      name:
        type: string
    type: object
...
info:
  contact: {}
paths:
...
  /role:
    get:
      description: Get a list of Role.
      parameters:
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Role'
            type: array
      security:
      - ApiKeyAuth: []
      summary: Get all Role.
      tags:
      - Role
  /role/:
    post:
      description: Creating a new Role.
      parameters:
      - description: the body to create a new Role
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.roleInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      summary: Create New Role`.
      tags:
      - Role
  /role/{id}:
    delete:
      description: Delete a Role by id.
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: string
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: boolean
            type: object
      security:
      - ApiKeyAuth: []
      summary: Delete one Role.
      tags:
      - Role
    get:
      description: Get an Role by id.
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      summary: Get Role.
      tags:
      - Role
    patch:
      description: Update Role by id.
      parameters:
      - description: Role id
        in: path
        name: id
        required: true
        type: string
      - description: the body to update age rating category
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/controllers.roleInput'
      - default: Bearer
        description: Insert your access token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Role'
      security:
      - ApiKeyAuth: []
      summary: Update Role.
      tags:
      - Role
...
swagger: "2.0"